+++ app/src/main/java/app/airsignal/weather/view/perm/FirstLocCheckDialog.kt
@@ -1,20 +1,22 @@
-package app.airsignal.weather.view
+package app.airsignal.weather.view.perm
 
 import android.annotation.SuppressLint
 import android.app.Activity
 import android.app.Dialog
+import android.content.Intent
+import android.net.Uri
 import android.os.*
+import android.provider.Settings
 import android.util.DisplayMetrics
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.view.animation.AnimationUtils
 import android.widget.*
 import androidx.appcompat.widget.AppCompatButton
 import androidx.core.content.res.ResourcesCompat
 import androidx.fragment.app.FragmentManager
 import app.airsignal.weather.R
-import app.airsignal.weather.util.RequestPermissionsUtil
+import app.airsignal.weather.util.`object`.GetAppInfo
 import app.airsignal.weather.util.`object`.SetAppInfo
 import com.google.android.material.bottomsheet.BottomSheetBehavior
 import com.google.android.material.bottomsheet.BottomSheetDialog
@@ -26,45 +28,60 @@ import kotlinx.coroutines.*
  * @author : Lee Jae Young
  * @since : 2023-04-11 오전 11:53
  **/
-class LocPermCautionDialog(
+class FirstLocCheckDialog(
     mActivity: Activity,
-    private val fm: FragmentManager, private val tagId: String?,
+    private val fm: FragmentManager, private val tagId: String?
 ) : BottomSheetDialogFragment() {
     private val activity = mActivity
+    private val perm = RequestPermissionsUtil(activity)
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        return inflater.inflate(R.layout.dialog_loc_perm_caution, container, false)
+        return inflater.inflate(R.layout.dialog_first_perm, container, false)
     }
 
     @SuppressLint("NotifyDataSetChanged")
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val okBtn = view.findViewById<AppCompatButton>(R.id.permCautionBtn)
-        val locIcon = view.findViewById<ImageView>(R.id.permCautionImg)
-        val locShadow = view.findViewById<ImageView>(R.id.permCautionImgShadow)
+        val apply = view.findViewById<AppCompatButton>(R.id.firstPermApplyBtn)
+        val cancel = view.findViewById<AppCompatButton>(R.id.firstPermCancelBtn)
 
-        // 권한 재요청 아이콘 애니메이션 적용
-        locIcon.animation =
-            AnimationUtils.loadAnimation(context,R.anim.loc_perm_caution_icon_anim).apply {
-                start()
+        apply.setOnClickListener {
+            if (!perm.isLocationPermitted()) {  // 위치 권한 허용?
+                if (perm.isShouldShowRequestPermissionRationale(
+                        activity,
+                        android.Manifest.permission.ACCESS_FINE_LOCATION
+                    )   // 권한 거부가 2번 이하?
+                ) {
+                    when (GetAppInfo.getInitLocPermission(activity)) { // 위치 권한 요청이 처음?
+                        "" -> {
+                            SetAppInfo.setInitLocPermission(activity, "Second")
+                            perm.requestLocation()
+                        }
+                        "Second" -> {
+                            LocPermCautionDialog(
+                                activity,
+                                fm,
+                                BottomSheetDialogFragment().tag
+                            )
+                                .show()
+                        }
+                    }
+                } else {
+                    val intent = Intent(Settings.ACTION_APPLICATION_DETAILS_SETTINGS)
+                    val uri: Uri = Uri.fromParts("package", activity.packageName, null)
+                    intent.data = uri
+                    startActivity(intent)
+                }
             }
-
-        // 권한 재요청 그림자 애니메이션 적용
-        locShadow.animation =
-            AnimationUtils.loadAnimation(context,R.anim.loc_perm_caution_shadow_anim).apply {
-                start()
-            }
-
-        // 확인 버튼 클릭
-        okBtn.setOnClickListener {
             dismissNow()
-            SetAppInfo.setInitLocPermission(activity, "Done")
-            RequestPermissionsUtil(activity).requestLocation()
+        }
+        cancel.setOnClickListener {
+            dismissNow()
         }
     }
 
@@ -76,7 +93,7 @@ class LocPermCautionDialog(
         dialog.setOnShowListener { dialogInterface ->
             val bottomSheetDialog = dialogInterface as BottomSheetDialog
             bottomSheetDialog.behavior.isDraggable = false
-            setupRatio(bottomSheetDialog, 60)
+            setupRatio(bottomSheetDialog, 75)
         }
         dialog.window?.attributes?.windowAnimations = R.style.DialogAnimationBottom
 
@@ -85,7 +102,7 @@ class LocPermCautionDialog(
 
     // 레이아웃 노출
     fun show() {
-        LocPermCautionDialog(activity, fm, tagId).showNow(fm, tagId)
+        FirstLocCheckDialog(activity, fm, tagId).showNow(fm, tagId)
     }
 
     // 바텀 다이얼로그 세팅
@@ -97,7 +114,8 @@ class LocPermCautionDialog(
         layoutParams.height = getBottomSheetDialogDefaultHeight(ratio)
         bottomSheet.layoutParams = layoutParams
         behavior.state = BottomSheetBehavior.STATE_EXPANDED
-        bottomSheet.background = ResourcesCompat.getDrawable(resources, R.drawable.loc_perm_bg,null)
+        bottomSheet.background =
+            ResourcesCompat.getDrawable(resources, R.drawable.loc_perm_bg, null)
     }
 
     // 바텀 다이얼로그 비율설정
